Fixed anchordetailscomponent call issues. Refactor the name of the exit and help popups and moved them into their own folder with the components file
